# api/openapi.yaml
openapi: 3.1.0
info:
  title: Dira Core API
  version: 0.1.0
  description: >-
    Socle de la plateforme Dira : identité, portefeuilles, paiements,
    notifications, véhicules et notes. Les verticales `dira-food-api`
    (livraison) et `dira-vtc-api` (courses) se posent dessus.


    ⚠️ Derrière la passerelle, ces routes sont servies à la RACINE
    (`/api/v1/...`), sans préfixe de verticale : l'authentification et le
    portefeuille sont les mêmes pour les deux, et les préfixer obligerait une
    application à choisir une verticale juste pour se connecter.


    ⚠️ Les routes `/internal/...` NE FIGURENT PAS ici. Elles portent les
    pouvoirs d'un service — débiter un portefeuille, ouvrir un compte — sont
    gardées par un secret partagé, et ne sont appelées par aucune application.
    Leur liste vit dans `internal/serviceapi/`, et un test refuse toute route
    qui s'y ajouterait sans être déclarée.


    **Erreurs.** Une seule enveloppe : `{"error":{"code","message"}}`, le
    message dans la langue de `Accept-Language`. Un `validation_failed`
    (422) ajoute `fields` — les clés JSON en cause — et, quand la valeur
    d'un champ n'est pas en cause, `reason` : `unknown_field` (une clé que
    la route ne connaît pas : elle est REFUSÉE, pas ignorée, pour qu'une
    application ne croie pas avoir enregistré ce qui a été jeté) ou
    `invalid_json`.


    **Téléphones.** Toujours en E.164 avec le `+` et l'indicatif ; les
...


...
    `POST /me/country/resolve`.
servers:
...
paths:
...
  /admin/wallets/{ownerID}:
    get:
      tags: [wallet]
      summary: One wallet, whoever owns it
      description: >-
        GÉNÉRIQUE : un propriétaire est un point de vente, un livreur ou un
        chauffeur — le socle ne fait pas la différence, et n'a pas à la faire.
      parameters: [{name: ownerID, in: path, required: true, schema: {type: string}}]
      responses:
        "200": {description: Wallet}
  /admin/wallets/{ownerID}/purchase:
    post:
      tags: [wallet]
      summary: Start a token purchase for an owner
      description: Initie un vrai paiement, à honorer.
      parameters: [{name: ownerID, in: path, required: true, schema: {type: string}}]
      responses:
        "201": {description: Payment started}
  /admin/wallets/{ownerID}/credit:
    post:
      tags: [wallet]
      summary: Grant tokens without payment
      description: >-
        ⚠️ SANS contrepartie financière — un geste commercial. Justificatif
        exigé, et l'écriture porte la trace de qui l'a accordé.
      parameters: [{name: ownerID, in: path, required: true, schema: {type: string}}]
      responses:
        "200": {description: Wallet}
  /admin/wallets/{ownerID}/transactions:
    get:
      tags: [wallet]
      summary: One wallet's ledger
      description: >-
        Le grand livre d'UN portefeuille, désigné par son propriétaire — la
        fiche d'un compte. Chaque mouvement dit son `unit` (`token` ou
        `xof`) : un relevé n'additionne pas des jetons et des francs.
      parameters:
        - {name: ownerID, in: path, required: true, schema: {type: string}}
        - {name: cursor, in: query, schema: {type: string}}
        - {name: limit, in: query, schema: {type: integer}}
      responses:
        "200": {description: Transactions}
        "404": {description: wallet_not_found}
  /admin/wallets/{ownerID}/promo:
    post:
      tags: [wallet]
      summary: Grant a promotional credit (XOF) to a client
      description: >-
        Un geste commercial en ARGENT — dépensé avant l'argent réel du
        client, jamais confondu avec lui au grand livre (`promo_credit`).
        Justificatif exigé, journal d'audit. Le portefeuille est créé s'il
        n'existe pas.
      parameters: [{name: ownerID, in: path, required: true, schema: {type: string}}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount_xof, justification]
              properties:
                amount_xof: {type: integer, minimum: 1}
                justification: {type: string, minLength: 3, maxLength: 500}
      responses:
        "200": {description: Wallet}
  /admin/wallets:
    get:
      tags: [wallet]
      summary: Les portefeuilles du pays courant — quatre soldes et le titulaire
      parameters:
        - {name: type, in: query, schema: {type: string, enum: [client, driver, merchant]}}
        - {name: owner_id, in: query, schema: {type: string}}
        - {name: cursor, in: query, schema: {type: string}}
        - {name: limit, in: query, schema: {type: integer, maximum: 100}}
      responses:
        "200": {description: "Wallets — `balance` (jetons), `balance_xof`, `promo_xof`, `debt_xof`, `owner_name` pour un compte (une boutique reste un identifiant)"}
  /admin/ledger:
    get:
      tags: [wallet]
      summary: Les mouvements de tous les portefeuilles du pays courant, les derniers d'abord
      description: >-
        Chaque ligne dit son UNITÉ (`token` ou `xof`), son motif, ce qu'elle
        paie (`ref_kind`, `ref_id`), son titulaire, et `ref` (la source d'un
        paiement, `debt_delta` d'un mouvement de dette).
      parameters:
        - {name: wallet_id, in: query, schema: {type: string}}
        - {name: owner_type, in: query, schema: {type: string, enum: [client, driver, merchant]}}
        - {name: unit, in: query, schema: {type: string, enum: [token, xof]}}
        - {name: reason, in: query, schema: {type: string}}
        - {name: ref_kind, in: query, schema: {type: string}}
        - {name: cursor, in: query, schema: {type: string}}
        - {name: limit, in: query, schema: {type: integer, maximum: 100}}
      responses:
        "200": {description: Ledger rows}
  /admin/wallets/{ownerID}/settle-debt:
    post:
      tags: [wallet]
      summary: Une dette réglée à l'agence (espèces, mobile money), hors solde
      description: >-
        `debt_xof` d'un portefeuille : la commission d'une course en espèces
        que le solde n'a pas pu payer, un paiement refusé à la livraison.
        Remboursée d'office sur le prochain crédit en argent ; ici, ce qui a
        été réglé en main propre. Jamais plus que la dette (422).
      parameters: [{name: ownerID, in: path, required: true, schema: {type: string}}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount_xof]
              properties:
                amount_xof: {type: integer, minimum: 1}
                note: {type: string, maxLength: 300}
      responses:
        "200": {description: Wallet — `debt_xof` mis à jour}
...
components:
...
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
security:
  - bearerAuth: []